### 2015-06-16 #17 @fmichonneau
rotl (Open Tree of Life)
* 1. What does this package do? (explain in 50 words or less)

It interacts with the Open Tree of Life API (http://opentreeoflife.org/)

* 2. Paste the full DESCRIPTION file inside a code block (bounded by ``` on either end).

```
Package: rotl
Title: Interface to the Open Tree of Life API
Version: 0.1
  person("Francois", "Michonneau", role=c("aut", "cre"),
         email="francois.michonneau@gmail.com"),
  person("Joseph", "Brown", role="aut"),
  person("David", "Winter", role="aut"))
Description: An interface to the Open Tree of Life API to retrieve phylogenetic
   trees, information about studies used to assemble the synthetic tree, and
   utilities to match taxonomic names to Open Tree identifiers. The Open Tree of
   Life aims at assembling a comprehensive phylogenetic tree for all named
   species.
    httr,
    rncl,
    ape
License: BSD_2_clause + file LICENSE
    RNeXML
```

* 3. URL for the package (the development repository, not a stylized html page)

https://github.com/fmichonneau/rotl

* 4. What data source(s) does it work with (if applicable)?

http://api.opentreeoflife.org/

* 5. Who is the target audience?

Scientists who want to use phylogenies

* 6. Are there other R packages that accomplish the same thing? If so, what is different about yours?

No.

* 7. Check the box next to each policy below, confirming that you agree. These are mandatory.
* [X] This package does not violate the Terms of Service of any service it interacts with.
* [X] The repository has continuous integration with Travis and/or another service
* [X] The package contains a vignette
* [X] The package contains a reasonably complete readme with devtools install instructions
* [X] The package contains unit tests
* [X] The package only exports functions to the NAMESPACE that are intended for end users
* 8. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.
* [ ] Are there any package dependencies not on CRAN?
* [X] Does the package have a CRAN accepted license?
* [ ] Did `devtools::check()` produce any errors or warnings? If so paste them below.
* 9. Please add explanations below for any exceptions to the above:
* 10. If this is a resubmission following rejection, please explain the change in cirucmstances.
