+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: rotl
 Title: Interface to the Open Tree of Life API
-Version: 0.1
+Version: 0.2
 Authors@R: c(
   person("Francois", "Michonneau", role=c("aut", "cre"),
          email="francois.michonneau@gmail.com"),
