### 2015-06-17 c65c877
list_synonyms_match_names now uses method synonyms
