+++ R/match_names.R
@@ -150,13 +150,13 @@ update_match_names <- function(response, row_number, taxon_name, ott_id,
 ##' returning an accepted synonym.
 ##'
 ##' To list synonyms for a given taxonomic name, you need to provide
-##' the object resulting from a call to the tnrs_match_names function,
-##' as well as one of either the row number corresponding to the name
-##' in this object, the name itself (as used in the original query),
-##' or the ott_id listed for this name.
+##' the object resulting from a call to the \code{tnrs_match_names}
+##' function, as well as one of either the row number corresponding to
+##' the name in this object, the name itself (as used in the original
+##' query), or the ott_id listed for this name.
 ##'
 ##' @title List the synonyms for a given name
-##' @param response a data frame generated by the tnrs_match_names
+##' @param tax a data frame generated by the tnrs_match_names
 ##'     function
 ##' @param row_number the row number corresponding to the name for
 ##'     which to list the synonyms
@@ -164,12 +164,22 @@ update_match_names <- function(response, row_number, taxon_name, ott_id,
 ##'     which to list the synonyms
 ##' @param ott_id the ott id corresponding to the name for which to
 ##'     list the synonyms
+##' @param ... currently ignored
 ##' @return a list whose elements are all synomym names (as vectors of
 ##'     character) for the taxonomic names that match the query (the
 ##'     names of the elements of the list).
+##' @examples
+##' \dontrun{
+##'    echino <- tnrs_match_names(c("Diadema", "Acanthaster", "Fromia"))
+##'    ## These 3 calls are identical
+##'    synonyms(echino, taxon_name="Acanthaster")
+##'    synonyms(echino, row_number=2)
+##'    synonyms(echino, ott_id=337928)
+##' }
 ##' @export
 
-list_synonyms_match_names <- function(response, row_number, taxon_name, ott_id) {
+synonyms.match_names <- function(tax, row_number, taxon_name, ott_id, ...) {
+    response <- tax
     i <- check_args_match_names(response, row_number, taxon_name, ott_id)
 
     res <- attr(response, "original_response")
