+++ NAMESPACE
@@ -14,6 +14,7 @@ S3method(print,gol)
 S3method(print,tnrs_contexts)
 S3method(print,tol_summary)
 S3method(study_list,tol_summary)
+S3method(synonyms,match_names)
 S3method(synonyms,taxon_info)
 S3method(synth_sources,gol_node)
 S3method(tax_rank,gol_node)
@@ -29,7 +30,6 @@ export(get_tree_ids)
 export(gol_about)
 export(gol_node_info)
 export(inspect_match_names)
-export(list_synonyms_match_names)
 export(node_id)
 export(ott_id)
 export(ott_taxon_name)

[…]
