### 2015-06-17 8376b5b
update_match_names now uses update method
