+++ man/inspect_match_names.Rd
@@ -2,14 +2,14 @@
 % Please edit documentation in R/match_names.R
 \name{inspect_match_names}
 \alias{inspect_match_names}
-\alias{update_match_names}
+\alias{update.match_names}
 \title{Inspect and Update alternative matches for a name returned by
     tnrs_match_names}
 \usage{
 inspect_match_names(response, row_number, taxon_name, ott_id)
 
-update_match_names(response, row_number, taxon_name, ott_id, new_row_number,
-  new_ott_id)
+\method{update}{match_names}(object, row_number, taxon_name, ott_id,
+  new_row_number, new_ott_id, ...)
 }
 \arguments{
 \item{response}{a data frame generated by the tnrs_match_names
@@ -22,6 +22,8 @@ inspect}
 
 \item{ott_id}{the ott id corresponding to the name to inspect}
 
+\item{object}{an object created by \code{\link{tnrs_match_names}}}
+
 \item{new_row_number}{the row number in the output of
 inspect_match_names to replace the taxa specified by
 \code{row_number}, \code{taxon_name}, or \code{ott_id}.}
@@ -29,6 +31,8 @@ inspect_match_names to replace the taxa specified by
 \item{new_ott_id}{the ott id of the taxon to replace the taxa
 specified by \code{row_number}, \code{taxon_name}, or
 \code{ott_id}.}
+
+\item{...}{currently ignored}
 }
 \value{
 a data frame
@@ -58,8 +62,8 @@ in which the name to be replaced appear or its ott id.
 \dontrun{
    matched_names <- tnrs_match_names(c("holothuria", "diadema", "boletus"))
    inspect_match_names(matched_names, taxon_name="diadema")
-   new_matched_names <- update_match_names(matched_names, taxon_name="diadema",
-                                           new_ott_id = 631176)
+   new_matched_names <- update(matched_names, taxon_name="diadema",
+                               new_ott_id = 631176)
    new_matched_names
    }
 }
