+++ NAMESPACE
@@ -20,6 +20,7 @@ S3method(synth_sources,gol_node)
 S3method(tax_rank,gol_node)
 S3method(tax_rank,taxon_info)
 S3method(tax_rank,taxon_lica)
+S3method(update,match_names)
 export(candidate_for_synth)
 export(get_publication)
 export(get_study)
@@ -51,6 +52,5 @@ export(tol_about)
 export(tol_induced_subtree)
 export(tol_mrca)
 export(tol_subtree)
-export(update_match_names)
 import(ape)
 import(rncl)
