### 2015-06-17 a769587
inspect_match_names now uses method inspect
