+++ R/match_names.R
@@ -51,26 +51,29 @@ check_args_match_names <- function(response, row_number, taxon_name, ott_id) {
 ##' To update one of the name, you also need to provide the row number
 ##' in which the name to be replaced appear or its ott id.
 ##'
-##' @title Inspect and Update alternative matches for a name returned by
-##'     tnrs_match_names
-##' @param response a data frame generated by the tnrs_match_names
-##'     function
+##' @title Inspect and Update alternative matches for a name returned
+##'     by tnrs_match_names
+##' @param response an object generated by the
+##'     \code{\link{tnrs_match_names}} function
 ##' @param row_number the row number corresponding to the name to
 ##'     inspect
-##' @param taxon_name the taxon name corresponding to the name to inspect
+##' @param taxon_name the taxon name corresponding to the name to
+##'     inspect
 ##' @param ott_id the ott id corresponding to the name to inspect
+##' @param ... currently ignored
 ##' @return a data frame
 ##' @seealso \code{\link{tnrs_match_names}}
 ##' @examples
 ##'   \dontrun{
 ##'    matched_names <- tnrs_match_names(c("holothuria", "diadema", "boletus"))
-##'    inspect_match_names(matched_names, taxon_name="diadema")
+##'    inspect(matched_names, taxon_name="diadema")
 ##'    new_matched_names <- update(matched_names, taxon_name="diadema",
 ##'                                new_ott_id = 631176)
 ##'    new_matched_names
 ##'    }
 ##' @export
-inspect_match_names <- function(response, row_number, taxon_name, ott_id) {
+##' @rdname match_names
+inspect.match_names <- function(response, row_number, taxon_name, ott_id, ...) {
 
     i <- check_args_match_names(response, row_number, taxon_name, ott_id)
 
@@ -90,16 +93,19 @@ inspect_match_names <- function(response, row_number, taxon_name, ott_id) {
     summary_match
 }
 
+##' @export
+##' @rdname match_names
+inspect <- function(response, ...) UseMethod("inspect")
+
 ##' @param object an object created by \code{\link{tnrs_match_names}}
 ##' @param new_row_number the row number in the output of
-##'     inspect_match_names to replace the taxa specified by
+##'     \code{\link{inspect}} to replace the taxa specified by
 ##'     \code{row_number}, \code{taxon_name}, or \code{ott_id}.
 ##' @param new_ott_id the ott id of the taxon to replace the taxa
 ##'     specified by \code{row_number}, \code{taxon_name}, or
 ##'     \code{ott_id}.
-##' @param ... currently ignored
 ##' @export
-##' @rdname inspect_match_names
+##' @rdname match_names
 update.match_names <- function(object, row_number, taxon_name, ott_id,
                                new_row_number, new_ott_id, ...) {
 
