+++ man/tnrs_match_names.Rd
@@ -74,7 +74,7 @@ function.
 }
 }
 \seealso{
-\code{\link{inspect_match_names}},
+\code{\link{inspect.match_names}},
     \code{\link{update.match_names}},
     \code{\link{synonyms.match_names}}.
 }
