+++ NAMESPACE
@@ -3,6 +3,7 @@
 S3method(candidate_for_synth,study_meta)
 S3method(get_publication,study_meta)
 S3method(get_tree_ids,study_meta)
+S3method(inspect,match_names)
 S3method(node_id,gol_node)
 S3method(node_id,taxon_info)
 S3method(node_id,taxon_lica)
@@ -30,7 +31,7 @@ export(get_study_tree)
 export(get_tree_ids)
 export(gol_about)
 export(gol_node_info)
-export(inspect_match_names)
+export(inspect)
 export(node_id)
 export(ott_id)
 export(ott_taxon_name)
