### 2015-06-18 610b1ea
add build_vignette=TRUE to README
