+++ README.Rmd
@@ -36,7 +36,7 @@ Then you can intall `rotl` using:
 
 ```{r, eval=FALSE}
 library(devtools)
-install_github("fmichonneau/rotl")
+install_github("fmichonneau/rotl", build_vignette=TRUE)
 ```
 
 ## Vignettes
