+++ appveyor.yml
@@ -1,0 +1,39 @@
+environment:
+   R_CHECK_ARGS: --no-manual
+
+init:
+    ps: |
+        $ErrorActionPreference = "Stop"
+        Invoke-WebRequest http://raw.github.com/krlmlr/r-appveyor/master/scripts/appveyor-tool.ps1 -OutFile "..\appveyor-tool.ps1"
+        Import-Module '..\appveyor-tool.ps1'
+
+install:
+    ps: Bootstrap
+
+build_script:
+    - travis-tool.sh install_deps
+
+test_script:
+    - travis-tool.sh run_tests
+
+on_failure:
+    - travis-tool.sh dump_logs
+
+artifacts:
+    - path: '*.Rcheck\**\*.log'
+      name: Logs
+
+    - path: '*.Rcheck\**\*.out'
+      name: Logs
+
+    - path: '*.Rcheck\**\*.fail'
+      name: Logs
+
+    - path: '*.Rcheck\**\*.Rout'
+      name: Logs
+
+    - path: '\*_*.tar.gz'
+      name: Bits
+
+    - path: '\*_*.zip'
+      name: Bits
