### 2015-06-18 018fcb1
update appveyor badge URL
