### 2015-06-18 e19c8e4
use factory for match_names methods
