### 2015-06-18 662edd5
make method factory more general
