### 2015-06-18 56de552
new flags method
