+++ NAMESPACE
@@ -1,6 +1,7 @@
 # Generated by roxygen2 (4.1.1): do not edit by hand
 
 S3method(candidate_for_synth,study_meta)
+S3method(flags,match_names)
 S3method(get_publication,study_meta)
 S3method(get_tree_ids,study_meta)
 S3method(inspect,match_names)
@@ -23,6 +24,7 @@ S3method(tax_rank,taxon_info)
 S3method(tax_rank,taxon_lica)
 S3method(update,match_names)
 export(candidate_for_synth)
+export(flags)
 export(get_publication)
 export(get_study)
 export(get_study_meta)
