### 2015-06-18 4f4e1cc
method factory deals with names and list structure
