### 2015-06-19 db7b1bb
add tests for flags method
