### 2015-06-19 c91d254
fix typo in get_list_element function
