### 2015-06-19 e071dc4
tests objects for each context, tests for synonyms
