### 2015-06-19 3cfe9d7
simplify and all arguments only for the factory
