### 2015-06-19 051067b
add tests for node_id method for class match_names
