+++ tests/testthat/test-match_names.R
@@ -295,3 +295,44 @@ test_that("flags with ott id", {
     expect_true(inherits(flags_rsp[[1]], "character"))
     expect_equal(names(flags_rsp), tax_rsp[1])
 })
+
+
+############################################################################
+## node_id method                                                         ##
+############################################################################
+
+context("node_id method for class match_names")
+
+if (identical(Sys.getenv("NOT_CRAN"), "true")) {
+    tax_rsp <- c("Tyrannosaurus", "Helicoplacus", "Ctenocystis",
+                 "Holothuria", "Echinoidea")
+    rsp <- tnrs_match_names(tax_rsp)
+}
+
+test_that("node_id with no arguments", {
+    skip_on_cran()
+    expect_true(inherits(node_id(rsp), "integer"))
+    expect_equal(names(node_id(rsp)), tax_rsp)
+    expect_equal(node_id(rsp)["Holothuria"], 3315679)
+})
+
+test_that("node_id with row number", {
+    skip_on_cran()
+    expect_equal(length(node_id(rsp, 4)), 4)
+    expect_true(inherits(node_id(rsp, 4), "integer"))
+    expect_equivalent(node_id(rsp, 4)[1], 3315679)
+})
+
+test_that("node_id with taxon name", {
+    skip_on_cran()
+    expect_equal(length(node_id(rsp, taxon_name = "Holothuria")), 4)
+    expect_true(inherits(node_id(rsp, taxon_name = "Holothuria"), "integer"))
+    expect_equivalent(node_id(rsp, taxon_name = "Holothuria")[1], 3315679)
+})
+
+test_that("node_id with ott id", {
+    skip_on_cran()
+    expect_equal(length(node_id(rsp, ott_id=924443)), 4)
+    expect_true(inherits(node_id(rsp, ott_id=924443), "integer"))
+    expect_equivalent(node_id(rsp, ott_id=924443)[1], 3315679)
+})
