### 2015-06-19 8704436
make node_id and ott_id constructors more general
