### 2015-06-19 5ee2ec0
add tests for ott_id method for class match_names
