### 2015-06-19 a245d1d
ditch simplify and all for more consistency
