### 2015-06-19 32ae825
update tests to reflect match_names methods simplifications
