### 2015-06-19 d75ccc4
update documentation for new methods
