+++ R/match_names.R
@@ -165,7 +165,7 @@ get_list_element <- function(response, i, list_name) {
 
 match_names_method_factory <- function(list_name) {
 
-    function(tax, row_number, taxon_name, ott_id) {
+    function(tax, row_number, taxon_name, ott_id, ...) {
 
         response <- tax
         res <- attr(response, "original_response")
@@ -189,6 +189,48 @@ match_names_method_factory <- function(list_name) {
 
 }
 
+
+##' \code{rotl} provides a collection of functions that allows users
+##' to extract relevant information from an object generated by
+##' \code{\link{tnrs_match_names}} function.
+##'
+##' These methods optionally accept one of the arguments
+##' \code{row_number}, \code{taxon_name} or \code{ott_id} to retrieve
+##' the corresponding information for one of the matches in the object
+##' returned by the \code{\link{tnrs_match_names}} function.
+##'
+##' These methods will return the information for all the matches of
+##' the searched names, and not only the match currently used in data
+##' frame returned by \code{\link{tnrs_match_names}}. In other words,
+##' if \code{tax} is an object created by
+##' \code{\link{tnrs_match_names}}, \code{ott_id(tax)} and
+##' \code{tax$ott_id} are not identical (see the examples).
+##'
+##' @title \code{node_id}, \code{ott_id} and \code{flags} for
+##'     taxonomic names matched by \code{tnrs_match_names}
+##' @param tax an object returned by \code{\link{tnrs_match_names}}
+##' @param row_number the row number corresponding to the name for
+##'     which to list the synonyms
+##' @param taxon_name the taxon name corresponding to the name for
+##'     which to list the synonyms
+##' @param ott_id the ott id corresponding to the name for which to
+##'     list the synonyms
+##' @param ... currently ignored
+##' @return A list of the ott ids, node ids or flags for the taxonomic
+##'     names matched with \code{\link{tnrs_match_names}}, for either
+##'     one or all the names.
+##' @examples
+##' \dontrun{
+##'   rsp <- tnrs_match_names(c("Diadema", "Tyrannosaurus"))
+##'   rsp$ott_id    # ott id for match currently in use
+##'   ott_id(rsp)   # ott id for all possible taxonomic matches
+##'
+##'   ## flags() is useful for instance to determine if a taxon is extinct
+##'   flags(rsp, taxon_name="Tyrannosaurus")
+##' }
+##' @export
+##' @rdname match_names-methods
+ott_id.match_names <- match_names_method_factory("ot:ottId")
 
 ##' @export
 ##' @rdname match_names-methods
@@ -217,8 +259,8 @@ node_id.match_names <- match_names_method_factory("matched_node_id")
 ##' query), or the ott_id listed for this name.
 ##'
 ##' @title List the synonyms for a given name
-##' @param tax a data frame generated by the tnrs_match_names
-##'     function
+##' @param tax a data frame generated by the
+##'     \code{\link{tnrs_match_names}} function
 ##' @param row_number the row number corresponding to the name for
 ##'     which to list the synonyms
 ##' @param taxon_name the taxon name corresponding to the name for
