+++ man/gol_node_info.Rd
@@ -13,9 +13,9 @@ gol_node_info(node_id = NULL, ott_id = NULL, include_lineage = FALSE, ...)
 
 \method{tax_rank}{gol_node}(tax)
 
-\method{node_id}{gol_node}(tax)
+\method{node_id}{gol_node}(tax, ...)
 
-\method{ott_id}{gol_node}(tax)
+\method{ott_id}{gol_node}(tax, ...)
 
 synth_sources(tax)
 
