+++ man/synonyms.match_names.Rd
@@ -7,8 +7,8 @@
 \method{synonyms}{match_names}(tax, row_number, taxon_name, ott_id, ...)
 }
 \arguments{
-\item{tax}{a data frame generated by the tnrs_match_names
-function}
+\item{tax}{a data frame generated by the
+\code{\link{tnrs_match_names}} function}
 
 \item{row_number}{the row number corresponding to the name for
 which to list the synonyms}
