+++ man/match_names-methods.Rd
@@ -1,0 +1,66 @@
+% Generated by roxygen2 (4.1.1): do not edit by hand
+% Please edit documentation in R/match_names.R
+\name{ott_id.match_names}
+\alias{flags}
+\alias{flags.match_names}
+\alias{node_id.match_names}
+\alias{ott_id.match_names}
+\title{\code{node_id}, \code{ott_id} and \code{flags} for
+    taxonomic names matched by \code{tnrs_match_names}}
+\usage{
+\method{ott_id}{match_names}(tax, row_number, taxon_name, ott_id, ...)
+
+flags(tax, ...)
+
+\method{flags}{match_names}(tax, row_number, taxon_name, ott_id, ...)
+
+\method{node_id}{match_names}(tax, row_number, taxon_name, ott_id, ...)
+}
+\arguments{
+\item{tax}{an object returned by \code{\link{tnrs_match_names}}}
+
+\item{row_number}{the row number corresponding to the name for
+which to list the synonyms}
+
+\item{taxon_name}{the taxon name corresponding to the name for
+which to list the synonyms}
+
+\item{ott_id}{the ott id corresponding to the name for which to
+list the synonyms}
+
+\item{...}{currently ignored}
+}
+\value{
+A list of the ott ids, node ids or flags for the taxonomic
+    names matched with \code{\link{tnrs_match_names}}, for either
+    one or all the names.
+}
+\description{
+\code{rotl} provides a collection of functions that allows users
+to extract relevant information from an object generated by
+\code{\link{tnrs_match_names}} function.
+}
+\details{
+These methods optionally accept one of the arguments
+\code{row_number}, \code{taxon_name} or \code{ott_id} to retrieve
+the corresponding information for one of the matches in the object
+returned by the \code{\link{tnrs_match_names}} function.
+
+These methods will return the information for all the matches of
+the searched names, and not only the match currently used in data
+frame returned by \code{\link{tnrs_match_names}}. In other words,
+if \code{tax} is an object created by
+\code{\link{tnrs_match_names}}, \code{ott_id(tax)} and
+\code{tax$ott_id} are not identical (see the examples).
+}
+\examples{
+\dontrun{
+  rsp <- tnrs_match_names(c("Diadema", "Tyrannosaurus"))
+  rsp$ott_id    # ott id for match currently in use
+  ott_id(rsp)   # ott id for all possible taxonomic matches
+
+  ## flags() is useful for instance to determine if a taxon is extinct
+  flags(rsp, taxon_name="Tyrannosaurus")
+}
+}
+
