+++ NAMESPACE
@@ -6,9 +6,11 @@ S3method(get_publication,study_meta)
 S3method(get_tree_ids,study_meta)
 S3method(inspect,match_names)
 S3method(node_id,gol_node)
+S3method(node_id,match_names)
 S3method(node_id,taxon_info)
 S3method(node_id,taxon_lica)
 S3method(ott_id,gol_node)
+S3method(ott_id,match_names)
 S3method(ott_id,taxon_lica)
 S3method(ott_taxon_name,taxon_info)
 S3method(ott_taxon_name,taxon_lica)
