+++ R/tnrs.R
@@ -72,19 +72,9 @@ tnrs_match_names <- function(names = NULL, context_name = NULL,
                              include_dubious = include_dubious, ...)
 
     check_tnrs(res)
-    summary_match <- do.call("rbind", lapply(res$results, function(x) {
-        searchStr <- x$matches[[1]]$search_string
-        uniqNames <- x$matches[[1]]$unique_name
-        approxMatch <- x$matches[[1]]$is_approximate_match
-        ott_id <- x$matches[[1]]$'ot:ottId'
-        isSynonym <- x$matches[[1]]$is_synonym
-        isDeprecated <- x$matches[[1]]$is_deprecated
-        nMatch <- length(x$match)
-        c(searchStr, uniqNames, approxMatch, ott_id, nMatch, isSynonym, isDeprecated)
-    }))
-    summary_match <- data.frame(summary_match, stringsAsFactors=FALSE)
-    names(summary_match) <- c("search_string", "unique_name", "approximate_match",
-                              "ott_id", "number_matches", "is_synonym", "is_deprecated")
+    summary_match <- build_summary_match(res, res_id = seq_len(length(res[["results"]])),
+                                         match_id = 1)
+
     summary_match$search_string <- gsub("\\\\", "", summary_match$search_string)
     summary_match <- summary_match[match(tolower(names), summary_match$search_string), ]
     attr(summary_match, "original_order") <- as.numeric(rownames(summary_match))
@@ -101,6 +91,56 @@ check_tnrs <- function(req) {
     if (length(req$unmatched_name_ids)) {
         warning(paste(req$unmatched_name_ids, collapse=", "), " are not matched")
     }
+}
+
+tnrs_columns <- c("search_string" = "search_string",
+                  "unique_name" = "unique_name",
+                  "approximate_match" = "is_approximate_match",
+                  "ott_id" = "ot:ottId",
+                  #"node_id" = "matched_node_id",
+                  "is_synonym" = "is_synonym",
+                  "is_deprecated" = "is_deprecated")
+
+summary_row_factory <- function(res, res_id, match_id, columns = tnrs_columns) {
+    res_address <- res[["results"]][[res_id]][["matches"]][[match_id]]
+    ret <- sapply(columns, function(cols) res_address[[cols]])
+    nMatch <- length(res[["results"]][[res_id]][["matches"]])
+    c(ret, number_matches = nMatch)
+}
+
+build_summary_match <- function(res, res_id, match_id = NULL) {
+
+    if (length(res_id) > 1 &&
+       (!is.null(match_id) && length(match_id) > 1)) {
+        stop("Something is wrong. Please contact us.")
+    }
+
+    build_summary_row <- lapply(res_id, function(rid) {
+        if (is.null(match_id)) {
+            match_id <- seq_len(length(res[["results"]][[rid]][["matches"]]))
+        }
+        res <- lapply(match_id, function(mid) {
+            summary_row_factory(res, rid, mid)
+        })
+        if (identical(length(match_id), 1L)) {
+            unlist(res)
+        } else res
+    })
+
+    if (identical(length(res_id), 1L)) {
+        build_summary_row <- unlist(build_summary_row,
+                                    recursive = FALSE)
+    }
+
+    ## Needed if only 1 row returned
+    if (!inherits(build_summary_row, "list")) {
+        build_summary_row <- list(build_summary_row)
+    }
+
+    summary_match <- do.call("rbind", build_summary_row)
+    summary_match <- data.frame(summary_match, stringsAsFactors=FALSE)
+    names(summary_match) <- c(names(tnrs_columns), "number_matches")
+    summary_match
 }
 
 ##' This function returns a list of pre-defined taxonomic contexts
