### 2015-06-20 1426ff9
tweak size of plot in vignette
