### 2015-06-20 a98991d
add that pandoc is needed to build the vignette
