### 2015-06-20 0a6aee1
[fix #26] knitr needs to be 1.10.5+
