+++ DESCRIPTION
@@ -22,8 +22,8 @@ Imports:
     ape
 License: BSD_2_clause + file LICENSE
 Suggests:
-    knitr,
-    rmarkdown,
+    knitr (>= 1.10.5),
+    rmarkdown (>= 0.7),
     testthat,
     RNeXML
 VignetteBuilder: knitr
