### 2015-06-20 b5949ca
simplify phylo_from_otl
