### 2015-06-20 72149bb
[fix #28] implement tree_to_labels
function to parse labels from newick strings, so we don't have to rely on other
tools to just get labels. As a side effect, it allows users to get labels from
nodes with 1 descendant.
