### 2015-06-20 d88d852
[fix #30] match_names methods return only the match
