+++ R/match_names.R
@@ -150,7 +150,7 @@ get_list_element <- function(response, i, list_name) {
 
 match_names_method_factory <- function(list_name) {
 
-    function(tax, row_number, taxon_name, ott_id, ...) {
+    function(tax, row_number, taxon_name, ott_id, only_current = TRUE, ...) {
 
         response <- tax
         res <- attr(response, "original_response")
@@ -164,6 +164,16 @@ match_names_method_factory <- function(list_name) {
             names(ret) <- sapply(attr(response, "original_order"), function(i) {
                 get_list_element(res, i, "matched_name")[[1]]
             })
+            if (only_current)  {
+                ## ret is already in the correct order so we can use a sequence
+                ## to extract the correct element
+                ret <- mapply(function(x, i) {
+                    ret[[x]][i]
+                }, seq_along(ret), attr(response, "match_id"))
+                if (all(sapply(ret, length) == 1)) {
+                    ret <- unlist(ret, use.names = T)
+                }
+            }
         } else {
             i <- check_args_match_names(response, row_number, taxon_name, ott_id)
             ret <- get_list_element(res, i, list_name)
@@ -184,12 +194,10 @@ match_names_method_factory <- function(list_name) {
 ##' the corresponding information for one of the matches in the object
 ##' returned by the \code{\link{tnrs_match_names}} function.
 ##'
-##' These methods will return the information for all the matches of
-##' the searched names, and not only the match currently used in data
-##' frame returned by \code{\link{tnrs_match_names}}. In other words,
-##' if \code{tax} is an object created by
-##' \code{\link{tnrs_match_names}}, \code{ott_id(tax)} and
-##' \code{tax$ott_id} are not identical (see the examples).
+##' If these arguments are not provided, these methods can return
+##' information for the matches currently listed in the object
+##' returned by \code{\link{tnrs_match_names}} (the default) or all
+##' the matches (using \code{only_current = FALSE}).
 ##'
 ##' @title \code{node_id}, \code{ott_id} and \code{flags} for
 ##'     taxonomic names matched by \code{tnrs_match_names}
@@ -200,6 +208,10 @@ match_names_method_factory <- function(list_name) {
 ##'     which to list the synonyms
 ##' @param ott_id the ott id corresponding to the name for which to
 ##'     list the synonyms
+##' @param only_current logical (default \code{TRUE}), should the
+##'     results include data for all matched names, or only the one
+##'     listed in the object returned by
+##'     \code{\link{tnrs_match_names}}?
 ##' @param ... currently ignored
 ##' @return A list of the ott ids, node ids or flags for the taxonomic
 ##'     names matched with \code{\link{tnrs_match_names}}, for either
@@ -238,10 +250,13 @@ node_id.match_names <- match_names_method_factory("matched_node_id")
 ##' returning an accepted synonym.
 ##'
 ##' To list synonyms for a given taxonomic name, you need to provide
-##' the object resulting from a call to the \code{tnrs_match_names}
-##' function, as well as one of either the row number corresponding to
-##' the name in this object, the name itself (as used in the original
-##' query), or the ott_id listed for this name.
+##' the object resulting from a call to the
+##' \code{\link{tnrs_match_names}} function, as well as one of either
+##' the row number corresponding to the name in this object, the name
+##' itself (as used in the original query), or the ott_id listed for
+##' this name. Otherwise, the synonyms for all the currently matched
+##' names are returned. Using \code{only_current = FALSE} will return
+##' all synonyms for all possible matches.
 ##'
 ##' @title List the synonyms for a given name
 ##' @param tax a data frame generated by the
@@ -252,6 +267,10 @@ node_id.match_names <- match_names_method_factory("matched_node_id")
[…]
