+++ man/synonyms.match_names.Rd
@@ -4,7 +4,8 @@
 \alias{synonyms.match_names}
 \title{List the synonyms for a given name}
 \usage{
-\method{synonyms}{match_names}(tax, row_number, taxon_name, ott_id, ...)
+\method{synonyms}{match_names}(tax, row_number, taxon_name, ott_id,
+  only_current = TRUE, ...)
 }
 \arguments{
 \item{tax}{a data frame generated by the
@@ -18,6 +19,11 @@ which to list the synonyms}
 
 \item{ott_id}{the ott id corresponding to the name for which to
 list the synonyms}
+
+\item{only_current}{logical (default \code{TRUE}), should the
+results include data for all matched names, or only the one
+listed in the object returned by
+\code{\link{tnrs_match_names}}?}
 
 \item{...}{currently ignored}
 }
@@ -36,10 +42,13 @@ returning an accepted synonym.
 }
 \details{
 To list synonyms for a given taxonomic name, you need to provide
-the object resulting from a call to the \code{tnrs_match_names}
-function, as well as one of either the row number corresponding to
-the name in this object, the name itself (as used in the original
-query), or the ott_id listed for this name.
+the object resulting from a call to the
+\code{\link{tnrs_match_names}} function, as well as one of either
+the row number corresponding to the name in this object, the name
+itself (as used in the original query), or the ott_id listed for
+this name. Otherwise, the synonyms for all the currently matched
+names are returned. Using \code{only_current = FALSE} will return
+all synonyms for all possible matches.
 }
 \examples{
 \dontrun{
