### 2015-06-21 24bcf2b
print fxn for get_study_meta
