+++ R/studies.R
@@ -242,6 +242,11 @@ get_study_meta <- function(study_id, ...) {
     res
 }
 
+print.study_meta <- function(x, ...) {
+    cat("Metadata for OToL study ", attr(x, "study_id"), ". Contents:\n", sep="")
+    cat(paste0("  $nexml$", names(x$nexml)), sep="\n")
+}
+
 ##' @export
 ##' @rdname get_study_meta
 get_tree_ids <- function(sm) UseMethod("get_tree_ids")
