### 2015-06-21 26fe7e4
extra space in study_meta print fxn
