+++ DESCRIPTION
@@ -25,6 +25,7 @@ Suggests:
     knitr (>= 1.10.5),
     rmarkdown (>= 0.7),
     testthat,
-    RNeXML
+    RNeXML, 
+    rfishbase
 VignetteBuilder: knitr
 LazyData: true
