+++ tests/testthat/test-taxonomy.R
@@ -48,6 +48,7 @@ test_that("taxonomy_taxon ott_taxon_name method", {
 
 test_that("taxonomy_taxon node_id method", {
     skip_on_cran()
+    expect_warning(node_id(tax_info), "node_ids are not persistent")
     expect_equal(names(node_id(tax_info)), as.character(tid))
     expect_equal(unname(node_id(tax_info)), c(3315679, 3297625, 3312119))
 })
@@ -137,6 +138,7 @@ test_that("lica ott_taxon_name method", {
 
 test_that("lica node_id method", {
     skip_on_cran()
+    expect_warning(node_id(tax_lica), "node_ids are not persistent")
     expect_equal(node_id(tax_lica), 3940323)
 })
 
