### 2015-06-29 #17 @sckott
Nice package! Overall, this is great! Some suggestions for improvements detailed below, as well as some ideas to ponder (but not necessarily to change anything).

## DESCRIPTION file

CRAN will likely complain about some of the words in your `Title` and `Description` fields - any that are not _normal_ dictionary words should get single quotes around them.

## Installation

Seems pretty easy on all platforms, windows and linux installs pass.

## Imports

* You have `httr` and `jsonlite` as Imports in `DESCRIPTION` file, but they aren't listed in any `@import`. Was there a reason behind this? Even though `httr` and `jsonlite` are namespaced, those will still fail for users that don't have those pkgs installed.
* You import all functions of your dependencies. It would be cleaner to use `@importFrom` and import only those functions you need instead of all of them.

## Docs

* Some examples in this vignette http://dwinter.github.io/rotl-vignette/ don't work. And it is linked to from the README, so would be good if that vignette works. Looks like the package vignette doesn't have this issue though. At least the following is broken:

```r
tr_string <- get_study_tree(study="2655", tree="tree6182", "newick")
#>  Error in match.arg(object_format) : 'arg' should be one of “phylo”
```

## Tests

All tests pass.

## Examples

Could use more examples for each function. Many only have 1 example. Especially - man pages for grouped functions often don't have examples for most of the functions in the man file  - though maybe there is a reason for this?

## Function Documentation

* You have for `tnrs_match_names`  for the `...` parameter

> additional arguments to customize the API request (see rotl package documentation).

And same for other function man files. But going to the linked page doesn't talk about what the `...` actually does. Drilling down, looks like the `...` gets passd on to `httr` HTTP methods, correct?

* Just a minor comment - for documentation bits that are duplicated, you could put documentation in `.R` files in a `man-roxygen` folder in the root of your repo, then use `@template filename` to use the docs.

## Specific comments about the scripts and functions:

- `ott_id` vs. `node_id` - I see a potential addition here. I imagine some users could get these confused since they are both short numeric values, and often the same length?  The functions that accept these ids as parameters arguments could have in addition generic S3 methods that act on each of them appropriately. So if you had functions that created S3 classes for each of `ott_id` vs. `node_id`, then functions could act on those correctly given their class.  This is just brainstorming - don't feel compelled to do this, but something to think about at least.
- Fail behavior:
    - Below I wonder if this could fail better?  Does the TOL API check that parameter inputs are of the correct type?  If so, could pass those error messages along to use. If not, could check these yourself.  I think most functions behave this way. I realize this exact mistake is not super likely, but better to program defensively.

    ```r
    tol_subtree(ott_id = "adfasdf")
    #> Error: HTTP failure: 400
    ```

## Readme:

> First find ott ids for a set of names:

What is an `ott`?

Also, in readme travis badge noticed that [Travis builds for mashup branch are failing](https://travis-ci.org/fmichonneau/rotl/builds/67776745#L1638). Looks like perhaps `ape` needs to be loaded in the vignette. Not a big deal as not master.

## Code of conduct:

- Nice, thanks for adding this!

## Documenting package changes

* I'd encourage authors to use a `NEWS` file to track changes, either as changes are made, or before new releases to CRAN. e.g., with links to issues (https://github.com/hadley/httr/blob/master/NEWS.md).  In addition, do tag each release to CRAN, and paste in associated items in your `NEWS` file into a release for that tag, e.g., https://github.com/hadley/httr/releases/tag/v1.0.0
