### 2015-07-04 fmichonneau/rotl#35 @fmichonneau
fix @import and @importFrom
add importFrom for functions needed for external packages.
