### 2015-07-04 fmichonneau/rotl#38 @fmichonneau
improve fail behavior
e.g.,

```r
tol_subtree(ott_id = "adfasdf")
```
