### 2015-07-04 4015f4e
[fix #34] add single quotes in DESCRIPTION
