### 2015-07-04 0bca62b
[fix #35] use importFrom insted of import
as a side effect, it seems that we don't need to import anything from ape
directly. The dependency on rncl is enough to provide access to the plotting
functions in ape apparently.
