+++ R/rotl-package.R
@@ -40,6 +40,4 @@
 ##'
 ##' @name rotl
 ##' @docType package
-##' @import ape
-##' @import rncl
 NULL
