+++ NAMESPACE
@@ -58,4 +58,10 @@ export(tol_induced_subtree)
 export(tol_mrca)
 export(tol_subtree)
 import(ape)
-import(rncl)
+importFrom(httr,GET)
+importFrom(httr,POST)
+importFrom(httr,content)
+importFrom(jsonlite,fromJSON)
+importFrom(jsonlite,toJSON)
+importFrom(jsonlite,unbox)
+importFrom(rncl,read_newick_phylo)
