### 2015-07-04 c307eac
error message if file given but file type is not
