### 2015-07-04 5f6b073
run examples on travis
