+++ .travis.yml
@@ -2,6 +2,8 @@ language: r
 
 sudo: required
 
+r_check_args: --as-cran --run-dontrun
+
 r_github_packages:
   - jimhester/covr
 
