### 2015-07-04 c29aeda
[fix #36] add more examples in documentation
