+++ man/get_study_meta.Rd
@@ -62,7 +62,9 @@ helper functions exists: \itemize{
 \examples{
 \dontrun{
 req <- get_study_meta("pg_719")
-req$nexml$`^ot:studyPublication`
+get_tree_ids(req)
+candidate_for_synth(req)
+get_publication(req)
 }
 }
 

[…]
