+++ NAMESPACE
@@ -57,7 +57,6 @@ export(tol_about)
 export(tol_induced_subtree)
 export(tol_mrca)
 export(tol_subtree)
-import(ape)
 importFrom(httr,GET)
 importFrom(httr,POST)
 importFrom(httr,content)
