### 2015-07-04 b64ff94
[fix #37] better doc of ... for call customization
