### 2015-07-04 06a2c6f
[fix #38] forgot to check_numeric on tol_* args
