+++ R/api-tol.R
@@ -18,14 +18,11 @@
     if (is.null(ott_ids) && is.null(node_ids)) {
         stop("Must supply at least one of \'ott_ids\' or \'node_ids\'.")
     }
-    ## It doesn't really make sense to enforce character here.
-    ## if (!is.null(ott_ids) && !is.character(ott_ids)) {
-    ##     stop("\'ott_ids\' must be of class \'character\'.")
-    ## }
-    ## if (!is.null(node_ids) && !is.character(node_ids)) {
-    ##     stop("\'ott_ids\' must be of class \'character\'.")
-    ## }
+
     if (is.null(node_ids) && !is.null(ott_ids)) {
+        if (!all(sapply(ott_ids, check_numeric))) {
+            stop("all ", sQuote("ott_ids"), " must look like numbers.")
+        }
         q <- list(ott_ids = ott_ids)
     }
     if (!is.null(node_ids) && is.null(ott_ids)) {
@@ -55,9 +52,13 @@
         warning("\'tree_id\' is currently ignored as the OTL only supports one version of tree of life.")
     }
     if (!is.null(ott_id)) {
+        if (length(ott_id) > 1) stop("A subtree can only be inferred from a single ", sQuote("ott_id"), ".")
+        if (!check_numeric(ott_id)) stop(sQuote("ott_id"), " needs to look like a number.")
         q <- list(ott_id = jsonlite::unbox(ott_id))
     }
     if (!is.null(node_id)) {
+        if (length(node_id) > 1) stop("A subtree can only be inferred from a single ", sQuote("node_id"), ".")
+        if (!check_numeric(node_id)) stop(sQuote("node_id"), " needs to look like a number.")
         q <- list(node_id = jsonlite::unbox(node_id))
     }
     res <- otl_POST(path="tree_of_life/subtree", body=q, ...)
@@ -77,10 +78,16 @@
         if (any(is.na(ott_ids))) {
         	stop("NAs are not allowed for argument \'ott_ids\'")
         }
+        if (!all(sapply(ott_ids, check_numeric))) {
+            stop(sQuote("ott_ids"), " must look like numbers")
+        }
     }
     if (!is.null(node_ids)) {
         if (any(is.na(node_ids))) {
         	stop("NAs are not allowed for argument \'node_ids\'")
+        }
+        if (!all(sapply(node_ids, check_numeric))) {
+            stop(sQuote("node_ids"), " must look like numbers")
         }
     }
 
