### 2015-07-04 9f2c71a
make travis stricter, attempt to remove NOTE
