+++ .travis.yml
@@ -2,6 +2,10 @@ language: r
 
 sudo: required
 
+warnings_are_errors: true
+
+before_install: echo "options(repos=c(CRAN = 'http://cran.rstudio.com'))" > ~/.Rprofile
+
 r_check_args: --as-cran --run-dontrun
 
 r_github_packages:
