### 2015-07-04 ropensci/rotl#41 @fmichonneau
drop all references to node_id(s)
as they will be removed from the API  in the future (see OpenTreeOfLife/treemachine#183, OpenTreeOfLife/taxomachine#94), are confusing and unreliable.
