### 2015-07-04 ad477ae
[fix #41] users can't use node_ids to call the API
