+++ tests/testthat/test-api-tol.R
@@ -16,32 +16,18 @@ test_that("study_list is logical for .tol_about", {
 
 test_that("neither ott_ids nor node_ids are NULL for .tol_mrca", {
     skip_on_cran()
-    expect_error(.tol_mrca(NULL, NULL),
+    expect_error(.tol_mrca(NULL),
                  "Must supply")
 })
-
-## test_that("ott_ids is character for .tol_mrca",
-##           expect_error(.tol_mrca(TRUE, NULL),
-##                        "character"))
-
-## test_that("node_ids is character for .tol_mrca",
-##           expect_error(.tol_mrca(NULL, TRUE),
-##                        "character"))
 
 ############################################################################
 ## .tol_subtree                                                           ##
 ############################################################################
 
-test_that("neither node_id nor ott_id are NULL", {
+test_that("ott_id is not NULL", {
     skip_on_cran()
-    expect_error(.tol_subtree(node_id = NULL, tree_id = NULL),
-                 "Must supply")
-})
-
-test_that("node_id and ott_id are not both specified", {
-    skip_on_cran()
-    expect_error(.tol_subtree(node_id = 123, ott_id = 123),
-                 "Use only")
+    expect_error(.tol_subtree(ott_id = NULL, tree_id = NULL),
+                 "must be provided")
 })
 
 test_that("providing tree_id gives a warning", {
@@ -54,28 +40,14 @@ test_that("providing tree_id gives a warning", {
 ## .tol_induced_subtree                                                   ##
 ############################################################################
 
-test_that("neither node_ids nor ott_ids are NULL", {
+test_that("ott_ids is not NULL", {
     skip_on_cran()
-    expect_error(.tol_induced_subtree(NULL, NULL),
-                 "Must supply")
-})
-
-test_that("NAs are not accepted for node_ids", {
-    skip_on_cran()
-    expect_error(.tol_induced_subtree(node_ids = c(123, NA, 456)),
-                 "NAs are not allowed")
+    expect_error(.tol_induced_subtree(ott_ids = NULL),
+                 "must be provided")
 })
 
 test_that("NAs are not accepted for ott_ids", {
     skip_on_cran()
     expect_error(.tol_induced_subtree(ott_ids = c(123, NA, 456)),
                  "NAs are not allowed")
-})
-
-test_that("NAs are not accepted for ott_ids", {
-    skip_on_cran()
-    expect_error(.tol_induced_subtree(
-              node_ids = c(123, NA, 456),
-              ott_ids = c(123, NA, 456)),
-              "NAs are not allowed")
 })
