+++ R/api-tol.R
@@ -14,23 +14,16 @@
 
 ##' @importFrom httr content
 ## Get the MRCA of a set of nodes
-.tol_mrca <- function(ott_ids=NULL, node_ids=NULL, ...) {
-    if (is.null(ott_ids) && is.null(node_ids)) {
-        stop("Must supply at least one of \'ott_ids\' or \'node_ids\'.")
+.tol_mrca <- function(ott_ids=NULL, ...) {
+    if (is.null(ott_ids)) {
+        stop("Must supply at least some \'ott_ids\'.")
     }
 
-    if (is.null(node_ids) && !is.null(ott_ids)) {
+    if (!is.null(ott_ids)) {
         if (!all(sapply(ott_ids, check_numeric))) {
             stop("all ", sQuote("ott_ids"), " must look like numbers.")
         }
         q <- list(ott_ids = ott_ids)
-    }
-    if (!is.null(node_ids) && is.null(ott_ids)) {
-        q <- list(node_ids = node_ids)
-    }
-    if (!is.null(node_ids) && !is.null(ott_ids)) {
-        q <- list(ott_ids = ott_ids,
-                  node_ids = node_ids)
     }
     res <- otl_POST(path="tree_of_life/mrca", body=q, ...)
     cont <- httr::content(res)
@@ -41,12 +34,9 @@
 ##' @importFrom jsonlite unbox
 ##' @importFrom httr content
 ## Get a subtree from the OpenTree Tree of Life
-.tol_subtree <- function(node_id=NULL, ott_id=NULL, tree_id=NULL, ...) {
-    if (!is.null(node_id) && !is.null(ott_id)) {
-        stop("Use only node_id OR ott_id")
-    }
-    if (is.null(node_id) && is.null(ott_id)) {
-        stop("Must supply a \'node_id\' OR \'ott_id\'")
+.tol_subtree <- function(ott_id=NULL, tree_id=NULL, ...) {
+    if (is.null(ott_id)) {
+        stop("One \'ott_id\' must be provided")
     }
     if (!is.null(tree_id)) {
         warning("\'tree_id\' is currently ignored as the OTL only supports one version of tree of life.")
@@ -56,11 +46,6 @@
         if (!check_numeric(ott_id)) stop(sQuote("ott_id"), " needs to look like a number.")
         q <- list(ott_id = jsonlite::unbox(ott_id))
     }
-    if (!is.null(node_id)) {
-        if (length(node_id) > 1) stop("A subtree can only be inferred from a single ", sQuote("node_id"), ".")
-        if (!check_numeric(node_id)) stop(sQuote("node_id"), " needs to look like a number.")
-        q <- list(node_id = jsonlite::unbox(node_id))
-    }
     res <- otl_POST(path="tree_of_life/subtree", body=q, ...)
     cont <- httr::content(res)
     return(cont)
@@ -69,9 +54,9 @@
 
 ##' @importFrom httr content
 ## Get an induced subtree from the OpenTree Tree of Life from a set of nodes
-.tol_induced_subtree <- function(node_ids=NULL, ott_ids=NULL, ...) {
-    if (is.null(node_ids) && is.null(ott_ids)) {
-        stop("Must supply \'node_ids\' and/or \'ott_ids\'")
+.tol_induced_subtree <- function(ott_ids=NULL, ...) {
+    if (is.null(ott_ids)) {
+        stop("At least two valid ", sQuote("ott_ids"), " must be provided")
     }
 
     if (!is.null(ott_ids)) {
@@ -82,19 +67,9 @@
             stop(sQuote("ott_ids"), " must look like numbers")
         }
     }
-    if (!is.null(node_ids)) {
-        if (any(is.na(node_ids))) {
[…]
