+++ R/api-gol.R
@@ -40,27 +40,17 @@
 ##' @importFrom jsonlite unbox
 ##' @importFrom httr content
 ## Get summary information about a node in the Graph of Life
-.gol_node_info <- function(node_id=NULL, ott_id=NULL, include_lineage=FALSE, ...) {
-    if (!is.null(node_id) && !is.null(ott_id)) {
-        stop("Use only \'node_id\' OR \'ott_id\'")
-    }
-    if (is.null(node_id) && is.null(ott_id)) {
-        stop("Must supply a \'node_id\' OR \'ott_id\'")
-    }
+.gol_node_info <- function(ott_id=NULL, include_lineage=FALSE, ...) {
     if (!is.logical(include_lineage)) {
         stop("Argument \'include_lineage\' must be of class \"logical\"")
     }
-    if (!is.null(ott_id)) {
+    if (is.null(ott_id)) {
+        stop(sQuote("ott_id"), " needs to be provided")
+    } else {
         if (!check_numeric(ott_id)) {
             stop("Argument \'ott_id\' must look like a number.")
         }
         q <- list(ott_id=jsonlite::unbox(ott_id), include_lineage=jsonlite::unbox(include_lineage))
-    }
-    if (!is.null(node_id)) {
-        if (!check_numeric(node_id)) {
-            stop("Argument \'node_id\' must look like a number.")
-        }
-        q <- list(node_id=jsonlite::unbox(node_id), include_lineage=jsonlite::unbox(include_lineage))
     }
     res <- otl_POST(path="graph/node_info", body=q, ...)
     cont <- httr::content(res)

[…]
