### 2015-07-04 1f95e77
[fix #40] trees with 1 tip don't go through rncl
