+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: rotl
 Title: Interface to the 'Open Tree of Life' API
-Version: 0.2.4
+Version: 0.2.5
 Authors@R: c(
   person("Francois", "Michonneau", role=c("aut", "cre"),
          email="francois.michonneau@gmail.com"),
@@ -18,7 +18,7 @@ Depends:
 Imports:
     httr,
     jsonlite,
-    rncl
+    rncl (>= 0.2.4)
 License: BSD_2_clause + file LICENSE
 Suggests:
     knitr (>= 1.10.5),
