### 2015-07-05 9a8e35c
use github's rncl for appveyor
