+++ appveyor.yml
@@ -12,6 +12,7 @@ install:
 
 build_script:
     - travis-tool.sh install_deps
+    - travis-tool.sh install_github fmichonneau/rncl
 
 test_script:
     - travis-tool.sh run_tests
