### 2015-07-05 e010e88
[fix #43] clarify what ott ids are in README
