### 2015-07-05 #17 @fmichonneau
Thanks so much for this review @sckott. It's very useful and informative!

I created issues for most major points and referenced them below so you can see how I addressed your comments by looking at the associated commits/diffs.

The current remaining issue is fixing the "mash-up" vignette.

## DESCRIPTION file

> CRAN will likely complain about some of the words in your Title and Description fields - any that are not normal dictionary words should get single quotes around them

Thanks for pointing this out, I quoted 'Open Tree of Life' and 'Open Tree identifiers', see fmichonneau/rotl#34

## Imports

> It would be cleaner to use @importFrom and import only those functions you need instead of all of them.

Done. See fmichonneau/rotl#35

## Docs

> Some examples in this vignette http://dwinter.github.io/rotl-vignette/ don't work.

Yes, @dwinter is working on fixing it. We are keeping track of this at fmichonneau/rotl#23

## Examples

> Could use more examples for each function. Many only have 1 example.

I added a few more examples, see fmichonneau/rotl#36

> Especially - man pages for grouped functions often don't have examples for most of the functions in the man file - though maybe there is a reason for this?

Do you have any specific examples for this? I can think of `taxonomy-methods.Rd` that doesn't have any examples but individual methods have examples in other places. For instance, `synonyms` has examples in `synonyms.match_names.Rd` and `taxonomy_info.Rd`.

## Function documentation

> the linked page doesn't talk about what the ... actually does.

I tried to clarify the text in the documentation for the package, see fmichonneau/rotl#37. Let me know if it needs to be clarified further.

> for documentation bits that are duplicated, you could put documentation in .R files in a man-roxygen folder

Cool. I didn't know about this. I'll definitely use it the future!

## Specific comments about the scripts and functions:

> ott_id vs. node_id

I completely removed the ability for users to use `node_id`. Apparently, those are not meant to be used for anything but internally by the neo4j database. They will be removed from the API in the next version, so users can only interact with the API using `ott_id` now. See fmichonneau/rotl#41.

> Below I wonder if this could fail better?

Agreed. I had started doing this for other functions but I had missed a few. It should be fixed and fail more elegantly. See fmichonneau/rotl#38

```r
tol_subtree(ott_id = "fjklqem")
# Error in .tol_subtree(ott_id = ott_id, tree_id = tree_id, ...) :
#  ‘ott_id’ needs to look like a number.
```

## README

> What is an ott id?

Thanks for pointing it out. Tried to clarify: see fmichonneau/rotl#43

> that Travis builds for mashup branch are failing

This goes with fixing the other vignette. It should be resolved soon.

## NEWS file

That sounds good. I'll add a NEWS file when preparing for the CRAN release.
