### 2015-07-05 #17 @dwinter
Hi both, thanks @sckott for your time reviewing this and @fmichonneau for powering through these issues!

Just a note on the "mashup" vignette. It's going to be a little harder than I thought -- the tree I used in the earlier version has disappeared from OpenTree and a quick look at alternatives found less-than-great examples.

Do you have a time-line for CRAN submission / rOpenSci conclusion? I will try and work on this today, but might not have much time after that...
