### 2015-07-07 #17 @sckott
> Do you have any specific examples for this? I can think of taxonomy-methods.Rd that doesn't have any examples but individual methods have examples in other places...

Okay, I see what you mean, looks good.

WRT man-roxygen templates - make sure to put `man-roxygen` in `.Rbuildignore`

> the linked page doesn't talk about what the ... actually does.

fixes look good
