### 2015-07-09 35da0f8
Make objects returned by studies cleaner
* Adds a print functions for results of studies_find_*
* Unlist the list of properties returned by studies_properties so they
don't take us much screen real estate on printing
