### 2015-07-09 023a4ce
New mashup vignette for #23
Idea is just to show a simple workflow, and some of the
steps usually needed to join up tips (here as a `phylo4d`
object).

Need to work on the text still, but think this will be the basis of the
vignette
