+++ NAMESPACE
@@ -14,7 +14,9 @@ S3method(ott_id,match_names)
 S3method(ott_id,taxon_lica)
 S3method(ott_taxon_name,taxon_info)
 S3method(ott_taxon_name,taxon_lica)
+S3method(print,found_studies)
 S3method(print,gol)
+S3method(print,study_meta)
 S3method(print,tnrs_contexts)
 S3method(print,tol_summary)
 S3method(study_list,tol_summary)
