### 2015-07-09 2da653f
have to load pkg in pkg vignette :-\
