### 2015-07-09 4dc0441
clean up of mashup vignette text
