+++ vignettes/data_mashups.Rmd
@@ -20,9 +20,9 @@ Open Tree.
 
 Let's say you have a dataset where each row represents a measurement taken from
 one species, and your goal is to put these measurements in some phylogenetic
-context. . Here's a small example, the best available estimates of the
-mutation rate for a set of single-celled Eukaryotes:
-
+context. Here's a small example: the best estimate of the mutation rate for a
+set of unicellular Eukaryotes along with some other property of those species
+which might explain the mutation rate: 
 
 ```{r, data}
 csv_path <- system.file("extdata", "protist_mutation_rates.csv", package = "rotl")
@@ -30,20 +30,18 @@ mu <- read.csv(csv_path, stringsAsFactors=FALSE)
 mu
 ```
 
-To get started, we want to know that each of these species are known to the Open
-Tree project and what their unique ID is. We can use the Taxonomic Name
-resolution Service (`tnrs`) functions to do this. It can be useful to provide a
-taxonomic context for searches using `tnrs` function, so let's see if any of
-them apply to this group:
+If we want to get a tree for these species we need to start by finding the
+unique ID for each of these species in the Open Tree database. We can use the 
+Taxonomic Name Resolution Service (`tnrs`) functions to do this. Before we do
+that we should see if any of the taxonomic contexts, which can be used to narrow
+a search and avoid conflicts between different codes, apply to our group of species:
 
 ```{r, context}
 library(rotl)
 tnrs_contexts()
 ```
 
-
-No then. If all of our species fell into one of those groups we could specify
-that group to avoid conflicts between different taxonomic codes. As it is we can 
+Hmm, none of those groups contain all of our species. In this case we can 
 search using the `All life` context and the function `tnrs_match_names`:
 
 ```{r, match}
@@ -51,15 +49,16 @@ taxon_search <- tnrs_match_names(mu$species, context_name="All life")
 knitr::kable(taxon_search)
 ```
 
-So, all the species are known to Open Tree. Note, though, that one of the names
+Good, all of our  species are known to Open Tree. Note, though, that one of the names
 is a synonym. _Saccharomyces pombe_ is older name for what is now called 
-_Schizosaccharomyces pombe_. As the name suggests, the Taxonnomic Name
+_Schizosaccharomyces pombe_. As the name suggests, the Taxonomic Name
 Resolution Service is designed to deal with these problems (and similar ones
 like misspellings), but it is always a good idea to check the results of
-`tnrs_match_names` closely to ensure the results are what you expect.
+`tnrs_match_names` closely to ensure the results are what you expect. 
 
-Let's keep out original data in line with the Open Tree names and IDS by adding them to
-the `data.frame`
+In this case we have a good ID for each of our species so we can move on. Before 
+we do that, let's ensure we can match up our original data to the Open Tree
+names and IDs by adding them to our `data.frame`:
 
 ```{r, munge}
 mu$ott_name <- taxon_search$unique_name
@@ -69,20 +68,20 @@ mu$ott_id <- taxon_search$ott_id
 ##Find a tree with your taxa
 
 Now let's find a tree. There are two possible options here: we can search for
-published studies that include our taxa or we can use the synthetic tree from
-Open Tree. Let's start by searching for studies. Before we can do that, we can
-find the names of the various properties of studies or trees that can be used
-for searching:
+published studies that include our taxa or we can use the 'synthetic tree' from
+Open Tree. We can try both approaches.
 
 ###Published trees
 
+Before we can search for published studies or trees, we should check out the
+list of properties we can use to perform such searches:
 ```{r, properties}
[…]
