+++ vignettes/data_mashups.Rmd
@@ -147,13 +147,30 @@ the tree
 
 ```{r phylobase}
 library(phylobase)
-rownames(mu) <- mu$ott_name
-tree_data <- phylo4d(tr, mu[,2:4])
+mu_numeric <- mu[,c("mu", "pop.size", "genome.size")]
+rownames_numeric <- mu$ott_name
+tree_data <- phylo4d(tr, mu_numeric)
 ```
-And now we can plot the data and the tree togehter
+And now we can plot the data and the tree together
 
 
 ```{r,  fig.width=7, fig.height=5}
 plot(tree_data)
 ```
+
+##What next.
+
+The demonstration get's you to the point of visualizing your data in a
+phylogentic context. But there's a lot more you do with this sort of data in R.
+For instance, you could use packages like `ape`, `caper`, `phytools` and
+`mcmcGLMM` to perform phylogenetic comparative analyses of your data. You could
+gather more data on your species using pcakages that connect to
+trait databases like `rfishbase`, `AntWeb` or `rnpn` which provides data from
+the US National Phenology Network. You could also use `rentrez` to find genetic 
+data for each of your species, and use that data to generate branch lengths for
+the phylogeny. 
+
+
+
+
 
