### 2015-07-13 9b235fd
use CRAN version of covr
