+++ .travis.yml
@@ -8,9 +8,11 @@ before_install: echo "options(repos=c(CRAN = 'http://cran.rstudio.com'))" > ~/.R
 
 r_check_args: --as-cran --run-dontrun
 
+r_packages:
+  - covr
+
 r_github_packages:
   - fmichonneau/rncl
-  - jimhester/covr
 
 after_success:
    - Rscript -e 'library(covr); coveralls()'
