### 2015-07-13 #17 @fmichonneau
OK, I think we are good to go at this stage. I merged @dwinter's revamped vignette into master.

We may have to wait a few days for a CRAN release though, as I released the package rotl depends on (rncl) a few days ago. Unfortunately, in that release, while fixing a bug, I introduced stricter checks on the validity of the files. These checks made reading some OpenTree trees impossible. I resolved the issue in rncl but I don't know if it's wise to resubmit rncl to CRAN just 5 days following the previous submission. Do you have any advice on this?

Otherwise, if you are satisfied with the current status of rotl, let us know of the next step.

Thanks.

cc @josephwb
