### 2015-07-16 75ae57e
update README to get rncl from GH
