+++ README.Rmd
@@ -30,20 +30,22 @@ on CRAN.
 will also need to install that package.
 
 ```{r, eval=FALSE}
-install.packages(c("devtools", "rncl"))
+install.packages("devtools")
 ```
 
 Then you can install `rotl` using:
 
 ```{r, eval=FALSE}
 library(devtools)
-install_github("fmichonneau/rotl", build_vignette=TRUE)
+install_github("fmichonneau/rncl")
+install_github("fmichonneau/rotl", dependencies = TRUE, build_vignette=TRUE)
 ```
 
 To build the vignette, you will need to have pandoc 1.12.3 or higher
-installed. If you use RStudio, make sure you have v0.99 (as it comes with pandoc
-1.13.1); if you don't use RStudio, you can find pandoc for you operating system
-[here](http://pandoc.org/installing.html)
+installed. If you use
+[RStudio](https://www.rstudio.com/products/rstudio/download/), make sure you
+have v0.99 (as it comes with pandoc 1.13.1); if you don't use RStudio, you can
+find pandoc for your operating system [here](http://pandoc.org/installing.html)
 
 ## Vignettes
 
