### 2015-07-16 c6b95a9
update README with new vignette info
