+++ README.Rmd
@@ -50,11 +50,14 @@ find pandoc for your operating system [here](http://pandoc.org/installing.html)
 ## Vignettes
 
 There are two vignettes:
+
 - start by checking out the "How to use `rotl`?" by typing:
   `vignette("how-to-use-rotl", package="rotl")` after installing the
   package.
-- then explore in a little more detail how to combine `rotl` with other packages
-  with [this vignette](http://dwinter.github.io/rotl-vignette/)
+
+- then explore how you can use `rotl` with other packages to combine your data
+  with trees from the Open Tree of Life project by typing:
+  `vignette("data_mashups",package="rotl")`.
 
 ## Quick start
 
