### 2015-07-18 9ebc6d5
updates to reflect API changes
