+++ tests/testthat/test-gol.R
@@ -28,9 +28,9 @@ if (identical(Sys.getenv("NOT_CRAN"), "true")) {
 test_that("gol node info. If any of these tests break, it means that the doc for this function needs to be updated.", {
     skip_on_cran()
     expect_true(all(names(gol_info) %in%
-                      c("in_graph", "tree_id", "name","rank","ott_id", "num_tips",
+                      c("tree_id", "num_synth_tips", "name", "rank", "ott_id", "num_tips",
                         "tree_sources", "tax_source", "synth_sources",
-                        "node_id", "in_synth_tree", "num_synth_children")))
+                        "node_id", "in_synth_tree")))
     expect_true(inherits(gol_info, "gol_node"))
 })
 
