### 2015-07-18 466bd3e
fix typo in docs
