+++ man/gol_node_info.Rd
@@ -57,7 +57,7 @@ the node from the synthetic tree. Optional; default = FALSE.}
     trees are in the graph, but were not used in constructing the
     synthetic tree. Each source has:}
 
-    \itemize {
+    \itemize{
 
         \item {study_id} {The study identifier. Will typically include
         a prefix ("pg_" or "ot_").}
