### 2015-07-20 #17 @sckott
@fmichonneau sorry for the delay, just got back from 1 week vacation. i'll get to next steps soon...
