### 2015-07-20 #17 @sckott
@fmichonneau in terms of submitting `rncl` again to CRAN, they generally do get more annoyed the more often you submit a new version of a package, but it's worth a shot, and I think they are more lenient with people that maintain fewer packages (i.e., they don't like me)
