### 2015-07-20 #17 @fmichonneau
Thanks for the feedback and no worries about the delay. I'll try to submit
an update to `rncl` this evening and I'll keep you posted.

On Mon, Jul 20, 2015 at 3:12 PM, Scott Chamberlain <notifications@github.com
> wrote:

> @fmichonneau <https://github.com/fmichonneau> in terms of submitting rncl
> again to CRAN, they generally do get more annoyed the more often you submit
> a new version of a package, but it's worth a shot, and I think they are
> more lenient with people that maintain fewer packages (i.e., they don't
> like me)
>
