### 2015-07-20 #17 @sckott
@fmichonneau looks great to me, I marked as `approved`

Okay if we now move it into the ropensci organization account?  If so, i've created a team in our org account and invited you to it. I think after you accept you'll be able to transfer the repo to ropensci

Also, make sure to update github installation instructions and any repo links
