### 2015-07-20 #17 @fmichonneau
I transfered the repository to ropensci. I'll update all the links later this evening. Thanks!
